requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers
pillow
fastapi
uvicorn
python-multipart
gradio
pyyaml
pytest>=7.0.0
pytesseract
opencv-python
numpy
scikit-learn
tensorboard
